+++ CHANGELOG.md
@@ -2,8 +2,15 @@
 
 ## V0.4.0
 
-- [x] color-mix(srgb)
-- [x] color(srgb, srgb-linear, display-p3, prophoto-rgb, a98-rgb, rec2020, xyz, xyz-d50)
+CSS color level 4 & 5
+- [x] color space: srgb, srgb-linear, display-p3, prophoto-rgb, a98-rgb, rec2020, xyz, xyz-d50
+- [x] color-mix()
+- [x] color()
+- [x] relative color
+- [x] lab()
+- [x] lch()
+- [x] oklab()
+- [x] oklch()
 
 ## V0.3.0